Fix `RuntimeWarning: 'numba.runtests' found in sys.modules`
